Rewrite assembler_test_base.h
Simplify the code in preparation of move to LLVM prebuilt tools.

Bug: 147817558
Test: m test-art-host-gtest

+++ libartbase/base/common_art_test.cc
@@ -53,7 +53,7 @@ namespace art {
 
 using android::base::StringPrintf;
 
-ScratchDir::ScratchDir() {
+ScratchDir::ScratchDir(bool keep_files) : keep_files_(keep_files) {
   // ANDROID_DATA needs to be set
   CHECK_NE(static_cast<char*>(nullptr), getenv("ANDROID_DATA")) <<
       "Are you subclassing RuntimeTest?";
@@ -65,15 +65,17 @@ ScratchDir::ScratchDir() {
 }
 
 ScratchDir::~ScratchDir() {
-  // Recursively delete the directory and all its content.
-  nftw(path_.c_str(), [](const char* name, const struct stat*, int type, struct FTW *) {
-    if (type == FTW_F) {
-      unlink(name);
-    } else if (type == FTW_DP) {
-      rmdir(name);
-    }
-    return 0;
-  }, 256 /* max open file descriptors */, FTW_DEPTH);
+  if (!keep_files_) {
+    // Recursively delete the directory and all its content.
+    nftw(path_.c_str(), [](const char* name, const struct stat*, int type, struct FTW *) {
+      if (type == FTW_F) {
+        unlink(name);
+      } else if (type == FTW_DP) {
+        rmdir(name);
+      }
+      return 0;
+    }, 256 /* max open file descriptors */, FTW_DEPTH);
+  }
 }
 
 ScratchFile::ScratchFile() {
